adjested to new direstory structure addresed comments
