+++ pkg/ruby/gemspec/testdata/binary_json-2.3.0.gemspec
