+++ pkg/ruby/gemspec/testdata/json-2.3.0.gemspec
