+++ pkg/ruby/gemspec/testdata/matrix-0.3.1/local.gemspec
