+++ pkg/ruby/gemspec/testdata/net-ftp-0.1.1.gemspec
