+++ pkg/ruby/gemspec/testdata/rubygems-update.gemspec
